mathlib/matrix_fault_resolver.py
"""
Matrix Fault Resolver for Schwabot v0.3
Handles safe transitions between matrix states during faults
"""

import numpy as np
from dataclasses import dataclass
from typing import Dict, List, Tuple, Optional
from datetime import datetime
import json
import yaml
from pathlib import Path
import time
import random

@dataclass
class FaultState:
    """State of a matrix fault"""
    timestamp: str
    fault_type: str
    severity: float
    current_matrix: str
    target_matrix: str
    transition_started: bool = False
    transition_complete: bool = False
    retry_count: int = 0

class MatrixFaultResolver:
    """
    Implements safe matrix state transitions during faults
    """
    
    def __init__(self, config_path: str = "matrix_response_paths.yaml"):
        """Initialize matrix fault resolver"""
        self.config_path = config_path
        self.current_matrix = "default"
        self.fault_history: List[FaultState] = []
        self.load_config()
    
    def load_config(self):
        """Load configuration from YAML"""
        try:
            with open(self.config_path, 'r') as f:
                self.config = yaml.safe_load(f)
        except FileNotFoundError:
            self.config = {
                "retry_config": {
                    "base_delay": 1000,
                    "max_delay": 30000,
                    "backoff_factor": 2,
                    "jitter": True,
                    "jitter_factor": 0.1
                }
            }
    
    def calculate_retry_delay(self, retry_count: int) -> int:
        """
        Calculate retry delay with exponential backoff and jitter
        
        Args:
            retry_count: Number of retries attempted
            
        Returns:
            Delay in milliseconds
        """
        base_delay = self.config["retry_config"]["base_delay"]
        max_delay = self.config["retry_config"]["max_delay"]
        backoff_factor = self.config["retry_config"]["backoff_factor"]
        
        # Calculate exponential backoff
        delay = min(base_delay * (backoff_factor ** retry_count), max_delay)
        
        # Add jitter if enabled
        if self.config["retry_config"]["jitter"]:
            jitter_factor = self.config["retry_config"]["jitter_factor"]
            jitter = delay * jitter_factor
            delay = delay + random.uniform(-jitter, jitter)
        
        return int(delay)
    
    def get_fallback_strategy(self, fault_type: str) -> str:
        """
        Get fallback strategy for fault type
        
        Args:
            fault_type: Type of fault
            
        Returns:
            Fallback strategy name
        """
        return self.config.get(fault_type, {}).get("fallback_strategy", "emergency_hold")
    
    def handle_fault(self,
                    fault_type: str,
                    severity: float,
                    target_matrix: Optional[str] = None) -> FaultState:
        """
        Handle a matrix fault
        
        Args:
            fault_type: Type of fault
            severity: Fault severity
            target_matrix: Target matrix to transition to
            
        Returns:
            FaultState object
        """
        # Create fault state
        fault = FaultState(
            timestamp=datetime.now().isoformat(),
            fault_type=fault_type,
            severity=severity,
            current_matrix=self.current_matrix,
            target_matrix=target_matrix or self.get_fallback_strategy(fault_type)
        )
        
        # Add to history
        self.fault_history.append(fault)
        
        return fault
    
    def transition_matrix(self, fault: FaultState) -> bool:
        """
        Transition to new matrix state
        
        Args:
            fault: FaultState object
            
        Returns:
            True if transition successful
        """
        if fault.transition_complete:
            return True
        
        if not fault.transition_started:
            fault.transition_started = True
            fault.retry_count = 0
        
        # Check if we've exceeded max retries
        max_retries = self.config.get(fault.fault_type, {}).get("max_retries", 3)
        if fault.retry_count >= max_retries:
            return False
        
        # Calculate delay
        delay = self.calculate_retry_delay(fault.retry_count)
        
        # Attempt transition
        try:
            # Simulate transition (replace with actual matrix transition logic)
            time.sleep(delay / 1000)  # Convert ms to seconds
            
            # Update state
            self.current_matrix = fault.target_matrix
            fault.transition_complete = True
            return True
            
        except Exception as e:
            fault.retry_count += 1
            return False
    
    def get_fault_history(self) -> List[FaultState]:
        """
        Get fault history
        
        Returns:
            List of FaultState objects
        """
        return self.fault_history
    
    def get_current_matrix(self) -> str:
        """
        Get current matrix state
        
        Returns:
            Current matrix name
        """
        return self.current_matrix 